Fix flaky tests by centralizing state cleanup in conftest.py
Tests were flaky when run together because mutable global state
(favorite_detectors, favorite_extractors, favorite_localizers, and
settings persistence) was not consistently reset between tests.
Individual test files had fragile ad-hoc cleanup patterns (inline .pop()
calls, per-class autouse fixtures, setup_method/teardown_method) that
would leak state if assertions failed before the cleanup ran.

Changes:
- Rename reset_votes -> reset_state in conftest.py; add clearing of
  favorite_detectors, favorite_extractors, and favorite_localizers
- Add isolated_settings autouse fixture in conftest.py that redirects
  SETTINGS_PATH to a per-test tmp_path, preventing settings writes from
  leaking between tests via data/settings.json
- Remove 18 redundant per-file cleanup fixtures across 10 test files
  (clear_favorites, isolated_settings, reset_setting, setup/teardown)
- Remove fragile inline favorite_detectors.pop() calls (7 occurrences)
- Update CLAUDE.md with Test Isolation section documenting the pattern
